+++ mullvad-daemon/src/version_check.rs
@@ -47,7 +47,7 @@ const PLATFORM: &str = "windows";
 #[cfg(target_os = "android")]
 const PLATFORM: &str = "android";
 
-#[derive(Debug, Clone, Serialize, Deserialize, Eq, PartialEq)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 struct CachedAppVersionInfo {
     #[serde(flatten)]
     pub version_info: AppVersionInfo,
@@ -281,11 +281,20 @@ impl VersionUpdater {
             self.show_beta_releases || is_beta_version(),
         );
 
+        let wg_migration_threshold = if response.x_threshold_wg_default.is_nan() {
+            // If the value should for some strange reason be NaN then safe default to 0.0
+            0.0
+        } else {
+            // Make sure that the returned value is between 0% and 100%
+            response.x_threshold_wg_default.clamp(0.0, 1.0)
+        };
+
         AppVersionInfo {
             supported: response.supported,
             latest_stable: response.latest_stable.unwrap_or_else(|| "".to_owned()),
             latest_beta: response.latest_beta,
             suggested_upgrade,
+            wg_migration_threshold,
         }
     }
 
@@ -375,6 +384,7 @@ impl VersionUpdater {
                                     latest_stable: last_app_version_info.latest_stable,
                                     latest_beta: last_app_version_info.latest_beta,
                                     suggested_upgrade,
+                                    wg_migration_threshold: last_app_version_info.wg_migration_threshold,
                                 }).await;
                             }
                         }
@@ -468,6 +478,7 @@ fn dev_version_cache() -> AppVersionInfo {
         latest_stable: PRODUCT_VERSION.to_owned(),
         latest_beta: PRODUCT_VERSION.to_owned(),
         suggested_upgrade: None,
+        wg_migration_threshold: 1.0,
     }
 }